[Shree Marella] corrected data types of sub classes

+++ C0771612_MAD3004_W2020_MidTerm/C0771612_MAD3004_W2020_MidTerm/Bill.swift
@@ -20,9 +20,9 @@ class Bill : IDisplay
     var billIdentificationNumber : Int
     var billDate                 : Date
     var billType                 : BillType
-    var totalBillAmount          : Double
+    var totalBillAmount          : Float
     
-    init(_ billIdentificationNumber:Int,_ billDate:Date,_ billType:BillType,_  totalBillAmount:Double)
+    init(_ billIdentificationNumber:Int,_ billDate:Date,_ billType:BillType,_  totalBillAmount:Float)
     {
         self.billIdentificationNumber = billIdentificationNumber
         self.billDate                 = billDate
